requirements.txt
flask==1.1.2
flask_cors==3.0.10
gpt_2_simple==0.7.2
python-dotenv==0.17.0
tensorflow-gpu==1.15.5